Refactor legacy template code into dedicated package structure
Move all legacy template functionality from cmd/apps/init.go into a new
cmd/apps/legacytemplates package with clear separation of concerns:

- legacy.go: Manifest types and template loading
- prompts.go: User prompts for template selection and resources
- builders.go: Resource bindings and variables builders
- resources.go: Resource collection logic
- init.go: Template initialization workflow
- env_builder.go: Environment file generation (moved from cmd/apps)

Create cmd/apps/internal/yamlutil package to provide shared YAML utilities
and avoid import cycles between apps and legacytemplates packages.

Update cmd/apps/init.go and cmd/apps/import.go to use the new package structure.
