+++ cmd/apps/import_test.go
@@ -7,6 +7,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/databricks/cli/cmd/apps/internal/yamlutil"
 	"github.com/databricks/cli/libs/dyn"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -120,7 +121,7 @@ env: []`,
 
 			// Run function
 			appValue := tt.inputValue
-			filename, err := inlineAppConfigFile(&appValue)
+			filename, err := yamlutil.InlineAppConfigFile(&appValue)
 
 			if tt.expectError {
 				assert.Error(t, err)
@@ -173,7 +174,7 @@ func TestInlineAppConfigFileErrors(t *testing.T) {
 		require.NoError(t, err)
 
 		appValue := dyn.V(map[string]dyn.Value{"name": dyn.V("test")})
-		_, err = inlineAppConfigFile(&appValue)
+		_, err = yamlutil.InlineAppConfigFile(&appValue)
 		assert.Error(t, err)
 		assert.Contains(t, err.Error(), "failed to parse")
 	})
@@ -192,7 +193,7 @@ func TestInlineAppConfigFileErrors(t *testing.T) {
 		require.NoError(t, err)
 
 		appValue := dyn.V("not a map")
-		_, err = inlineAppConfigFile(&appValue)
+		_, err = yamlutil.InlineAppConfigFile(&appValue)
 		assert.Error(t, err)
 		assert.Contains(t, err.Error(), "app value is not a map")
 	})
@@ -231,7 +232,7 @@ func TestInlineAppConfigFileErrors(t *testing.T) {
 		}
 
 		appValue := dyn.V(map[string]dyn.Value{"name": dyn.V("test")})
-		_, err = inlineAppConfigFile(&appValue)
+		_, err = yamlutil.InlineAppConfigFile(&appValue)
 		assert.Error(t, err)
 		assert.Contains(t, err.Error(), "failed to read")
 	})
@@ -263,7 +264,7 @@ resources:
 			"description": dyn.V("existing description"),
 		})
 
-		filename, err := inlineAppConfigFile(&appValue)
+		filename, err := yamlutil.InlineAppConfigFile(&appValue)
 		require.NoError(t, err)
 		assert.Equal(t, "app.yml", filename)
 
@@ -306,7 +307,7 @@ env:
 		"name": dyn.V("test-app"),
 	})
 
-	filename, err := inlineAppConfigFile(&appValue)
+	filename, err := yamlutil.InlineAppConfigFile(&appValue)
 	require.NoError(t, err)
 	assert.Equal(t, "app.yml", filename)
