+++ cmd/apps/init.go
@@ -3,8 +3,7 @@ package apps
 import (
 	"bytes"
 	"context"
-	_ "embed"
-	"encoding/json"
+	_ "embed" // For embedding template files
 	"errors"
 	"fmt"
 	"os"
@@ -14,72 +13,23 @@ import (
 	"text/template"
 
 	"github.com/charmbracelet/huh"
+	"github.com/databricks/cli/cmd/apps/legacytemplates"
 	"github.com/databricks/cli/cmd/root"
 	"github.com/databricks/cli/libs/apps/features"
 	"github.com/databricks/cli/libs/apps/initializer"
 	"github.com/databricks/cli/libs/apps/prompt"
 	"github.com/databricks/cli/libs/cmdctx"
 	"github.com/databricks/cli/libs/cmdio"
-	"github.com/databricks/cli/libs/dyn"
-	"github.com/databricks/cli/libs/dyn/yamlsaver"
 	"github.com/databricks/cli/libs/git"
 	"github.com/databricks/cli/libs/log"
 	"github.com/spf13/cobra"
-	"go.yaml.in/yaml/v3"
 )
 
 const (
 	templatePathEnvVar = "DATABRICKS_APPKIT_TEMPLATE_PATH"
 	defaultTemplateURL = "https://github.com/databricks/appkit/tree/main/template"
-)
-
-//go:embed legacy-template/app-template-app-manifests.json
-var appTemplateManifestsJSON []byte
-
-//go:embed legacy-template/databricks-yml.tmpl
-var databricksYmlTemplate string
-
-//go:embed legacy-template/gitignore.tmpl
-var gitignoreTemplate string
-
-// resourceSpec represents a resource specification in the template manifest.
-type resourceSpec struct {
-	Name                string          `json:"name"`
-	Description         string          `json:"description"`
-	SQLWarehouseSpec    *map[string]any `json:"sql_warehouse_spec,omitempty"`
-	ExperimentSpec      *map[string]any `json:"experiment_spec,omitempty"`
-	ServingEndpointSpec *map[string]any `json:"serving_endpoint_spec,omitempty"`
-	DatabaseSpec        *map[string]any `json:"database_spec,omitempty"`
-	UCSecurableSpec     *map[string]any `json:"uc_securable_spec,omitempty"`
-}
-
-// manifest represents the manifest section of a template.
-type manifest struct {
-	Version       int            `json:"version"`
-	Name          string         `json:"name"`
-	Description   string         `json:"description"`
-	StartCommand  string         `json:"start_command,omitempty"`
-	ResourceSpecs []resourceSpec `json:"resource_specs,omitempty"`
-	UserAPIScopes []string       `json:"user_api_scopes,omitempty"`
-}
-
-// appTemplateManifest represents a single app template from the manifests JSON file.
-type appTemplateManifest struct {
-	Path     string   `json:"path"`
-	GitRepo  string   `json:"git_repo"`
-	Manifest manifest `json:"manifest"`
-}
-
-// appTemplateManifests holds all app templates.
-type appTemplateManifests struct {
-	Templates []appTemplateManifest `json:"appTemplateAppManifests"`
-}
-
-type templateType string
-
-const (
-	templateTypeAppKit templateType = "appkit"
-	templateTypeLegacy templateType = "legacy"
+	templateTypeAppKit = "appkit"
+	templateTypeLegacy = "legacy"
 )
 
 func newInitCmd() *cobra.Command {
@@ -499,367 +449,6 @@ func resolveTemplate(ctx context.Context, templatePath, branch string) (localPat
 	return tempDir, cleanup, nil
 }
 
-// loadLegacyTemplates loads the legacy app templates from the embedded JSON file.
-func loadLegacyTemplates() (*appTemplateManifests, error) {
-	var manifests appTemplateManifests
-	if err := json.Unmarshal(appTemplateManifestsJSON, &manifests); err != nil {
-		return nil, fmt.Errorf("failed to load app template manifests: %w", err)
-	}
-	return &manifests, nil
-}
-
-// findLegacyTemplateByPath finds a legacy template by its path identifier.
-// Returns nil if no matching template is found.
-func findLegacyTemplateByPath(manifests *appTemplateManifests, path string) *appTemplateManifest {
-	for i := range manifests.Templates {
-		if manifests.Templates[i].Path == path {
-			return &manifests.Templates[i]
-		}
-	}
-	return nil
-}
-
-// resourceSpecChecker is a function that checks if a resourceSpec has a specific field.
-type resourceSpecChecker func(*resourceSpec) bool
-
-// hasResourceSpec checks if a template requires a resource based on a spec checker.
-func hasResourceSpec(tmpl *appTemplateManifest, checker resourceSpecChecker) bool {
-	for _, spec := range tmpl.Manifest.ResourceSpecs {
-		if checker(&spec) {
-			return true
-		}
-	}
-	return false
-}
-
-func requiresSQLWarehouse(tmpl *appTemplateManifest) bool {
-	return hasResourceSpec(tmpl, func(s *resourceSpec) bool { return s.SQLWarehouseSpec != nil })
-}
-
-func requiresServingEndpoint(tmpl *appTemplateManifest) bool {
-	return hasResourceSpec(tmpl, func(s *resourceSpec) bool { return s.ServingEndpointSpec != nil })
-}
-
-func requiresExperiment(tmpl *appTemplateManifest) bool {
-	return hasResourceSpec(tmpl, func(s *resourceSpec) bool { return s.ExperimentSpec != nil })
-}
-
-func requiresDatabase(tmpl *appTemplateManifest) bool {
-	return hasResourceSpec(tmpl, func(s *resourceSpec) bool { return s.DatabaseSpec != nil })
-}
-
-func requiresUCVolume(tmpl *appTemplateManifest) bool {
-	return hasResourceSpec(tmpl, func(s *resourceSpec) bool { return s.UCSecurableSpec != nil })
-}
-
-// promptForTemplateType prompts the user to choose between AppKit and Legacy templates.
-func promptForTemplateType(ctx context.Context) (templateType, error) {
-	var choice string
-	options := []huh.Option[string]{
-		huh.NewOption("AppKit (TypeScript)", string(templateTypeAppKit)),
-		huh.NewOption("Legacy template", string(templateTypeLegacy)),
-	}
-
-	err := huh.NewSelect[string]().
-		Title("Select template type").
-		Options(options...).
-		Value(&choice).
-		WithTheme(prompt.AppkitTheme()).
-		Run()
-	if err != nil {
-		return "", err
-	}
-
-	prompt.PrintAnswered(ctx, "Template type", choice)
-	return templateType(choice), nil
-}
-
-// promptForLegacyTemplate prompts the user to select a legacy template.
-func promptForLegacyTemplate(ctx context.Context, manifests *appTemplateManifests) (*appTemplateManifest, error) {
-	options := make([]huh.Option[int], len(manifests.Templates))
-	for i := range manifests.Templates {
-		tmpl := &manifests.Templates[i]
-		label := tmpl.Path
-		if tmpl.Manifest.Name != "" {
-			label = tmpl.Path + " - " + tmpl.Manifest.Name
-			if tmpl.Manifest.Description != "" {
-				label = tmpl.Path + " - " + tmpl.Manifest.Name + " - " + tmpl.Manifest.Description
-			}
-		}
-		options[i] = huh.NewOption(label, i)
-	}
-
-	var selectedIdx int
-	err := huh.NewSelect[int]().
-		Title("Select a template").
-		Description("Choose from available templates").
-		Options(options...).
-		Value(&selectedIdx).
-		Height(15).
-		WithTheme(prompt.AppkitTheme()).
-		Run()
-	if err != nil {
-		return nil, err
-	}
-
-	selectedTemplate := &manifests.Templates[selectedIdx]
-	prompt.PrintAnswered(ctx, "Template", selectedTemplate.Path)
-	return selectedTemplate, nil
-}
-
-// resourceGetter defines how to get a resource value for a template.
-type resourceGetter struct {
-	checkRequired func(*appTemplateManifest) bool
-	promptFunc    func(context.Context) (string, error)
-	errorMessage  string
-}
-
-func writeGitignoreIfMissing(ctx context.Context, destDir string) error {
-	gitignorePath := filepath.Join(destDir, ".gitignore")
-
-	// Check if .gitignore already exists
-	if _, err := os.Stat(gitignorePath); err == nil {
-		// .gitignore already exists, skip
-		return nil
-	}
-
-	// Write the gitignore template
-	if err := os.WriteFile(gitignorePath, []byte(gitignoreTemplate), 0o644); err != nil {
-		return fmt.Errorf("failed to write .gitignore: %w", err)
-	}
-
-	cmdio.LogString(ctx, "✓ Created .gitignore")
-	return nil
-}
-
-// generateEnvFileForLegacyTemplate creates a .env file from app.yml with resource configuration.
-func generateEnvFileForLegacyTemplate(ctx context.Context, destDir, workspaceHost, profile, appName, warehouseID, servingEndpoint, experimentID, instanceName, databaseName, ucVolume string) error {
-	// Check if app.yml or app.yaml exists
-	var appYmlPath string
-	for _, filename := range []string{"app.yml", "app.yaml"} {
-		path := filepath.Join(destDir, filename)
-		if _, err := os.Stat(path); err == nil {
-			appYmlPath = path
-			break
-		}
-	}
-
-	if appYmlPath == "" {
-		// No app.yml found, skip .env generation
-		return nil
-	}
-
-	// Build resources map from collected parameters
-	// Resource names should match what's in databricks.yml
-	resources := make(map[string]string)
-
-	if warehouseID != "" {
-		resources[ResourceNameSQLWarehouse] = warehouseID
-	}
-	if servingEndpoint != "" {
-		resources[ResourceNameServingEndpoint] = servingEndpoint
-	}
-	if experimentID != "" {
-		resources[ResourceNameExperiment] = experimentID
-	}
-	if instanceName != "" {
-		resources[ResourceNameDatabase] = instanceName
-	}
-	if databaseName != "" {
-		resources[ResourceNameDatabaseName] = databaseName
-	}
-	if ucVolume != "" {
-		resources[ResourceNameUCVolume] = ucVolume
-	}
-
-	// Create EnvFileBuilder
-	builder, err := NewEnvFileBuilder(workspaceHost, profile, appName, appYmlPath, resources)
-	if err != nil {
-		return fmt.Errorf("failed to create env builder: %w", err)
-	}
-
-	// Write .env file
-	err = builder.WriteEnvFile(destDir)
-	if err != nil {
-		return fmt.Errorf("failed to write .env file: %w", err)
-	}
-
-	cmdio.LogString(ctx, "✓ Generated .env file from app.yml")
-	return nil
-}
-
-// resourceBinding represents a single resource binding in databricks.yml.
-type resourceBinding struct {
-	name        string
-	description string
-	lines       []string // Comment lines for the binding
-}
-
-// resourceBindingsBuilder constructs databricks.yml resource bindings.
-// Each add method appends a binding for a specific resource type.
-type resourceBindingsBuilder struct {
-	bindings []resourceBinding
-}
-
-func newResourceBindingsBuilder() *resourceBindingsBuilder {
-	return &resourceBindingsBuilder{bindings: make([]resourceBinding, 0)}
-}
-
-func (b *resourceBindingsBuilder) addWarehouse(warehouseID string) {
-	if warehouseID == "" {
-		return
-	}
-	b.bindings = append(b.bindings, resourceBinding{
-		name:        ResourceNameSQLWarehouse,
-		description: "SQL Warehouse for analytics",
-		lines: []string{
-			"          sql_warehouse:",
-			"            id: ${var.warehouse_id}",
-			"            permission: CAN_USE",
-		},
-	})
-}
-
-func (b *resourceBindingsBuilder) addServingEndpoint(endpoint string) {
-	if endpoint == "" {
-		return
-	}
-	b.bindings = append(b.bindings, resourceBinding{
-		name:        ResourceNameServingEndpoint,
-		description: "Model serving endpoint",
-		lines: []string{
-			"          serving_endpoint:",
-			"            name: ${var.serving_endpoint_name}",
-			"            permission: CAN_QUERY",
-		},
-	})
-}
-
-func (b *resourceBindingsBuilder) addExperiment(experimentID string) {
-	if experimentID == "" {
-		return
-	}
-	b.bindings = append(b.bindings, resourceBinding{
-		name:        ResourceNameExperiment,
-		description: "MLflow experiment",
-		lines: []string{
-			"          experiment:",
-			"            experiment_id: ${var.experiment_id}",
-			"            permission: CAN_MANAGE",
-		},
-	})
-}
-
-func (b *resourceBindingsBuilder) addDatabase(instanceName, databaseName string) {
-	if instanceName == "" || databaseName == "" {
-		return
-	}
-	b.bindings = append(b.bindings, resourceBinding{
-		name:        ResourceNameDatabase,
-		description: "Lakebase database",
-		lines: []string{
-			"          database:",
-			"            database_name: ${var.database_name}",
-			"            instance_name: ${var.instance_name}",
-			"            permission: CAN_CONNECT_AND_CREATE",
-		},
-	})
-}
-
-// build generates the resource bindings content for databricks.yml.
-func (b *resourceBindingsBuilder) build() string {
-	if len(b.bindings) == 0 {
-		return ""
-	}
-
-	var result []string
-	for _, binding := range b.bindings {
-		result = append(result, "        - name: "+binding.name)
-		result = append(result, "          description: "+binding.description)
-		result = append(result, binding.lines...)
-	}
-
-	return strings.Join(result, "\n")
-}
-
-// variablesBuilder builds bundle variables for databricks.yml.
-type variablesBuilder struct {
-	variables []struct {
-		name  string
-		value string
-	}
-}
-
-func newVariablesBuilder() *variablesBuilder {
-	return &variablesBuilder{}
-}
-
-func (b *variablesBuilder) addWarehouse(warehouseID string) {
-	if warehouseID != "" {
-		b.variables = append(b.variables, struct {
-			name  string
-			value string
-		}{"warehouse_id", warehouseID})
-	}
-}
-
-func (b *variablesBuilder) addServingEndpoint(endpoint string) {
-	if endpoint != "" {
-		b.variables = append(b.variables, struct {
-			name  string
-			value string
-		}{"serving_endpoint_name", endpoint})
-	}
-}
-
-func (b *variablesBuilder) addExperiment(experimentID string) {
-	if experimentID != "" {
-		b.variables = append(b.variables, struct {
-			name  string
-			value string
-		}{"experiment_id", experimentID})
-	}
-}
-
-func (b *variablesBuilder) addDatabase(instanceName, databaseName string) {
-	if databaseName != "" {
-		b.variables = append(b.variables, struct {
-			name  string
-			value string
-		}{"database_name", databaseName})
-	}
-	if instanceName != "" {
-		b.variables = append(b.variables, struct {
-			name  string
-			value string
-		}{"instance_name", instanceName})
-	}
-}
-
-func (b *variablesBuilder) addUCVolume(volume string) {
-	if volume != "" {
-		b.variables = append(b.variables, struct {
-			name  string
-			value string
-		}{"uc_volume", volume})
-	}
-}
-
-// build generates the variables content for databricks.yml.
-func (b *variablesBuilder) build() string {
-	if len(b.variables) == 0 {
-		return ""
-	}
-
-	var result []string
-	for _, v := range b.variables {
-		result = append(result, "  "+v.name+":")
-		result = append(result, "    default: "+v.value)
-	}
-
-	return strings.Join(result, "\n")
-}
-
 // deployRunConfig handles deploy and run mode determination.
 type deployRunConfig struct {
 	// From flags
@@ -892,538 +481,6 @@ func (c *deployRunConfig) resolve(ctx context.Context) (bool, prompt.RunMode, er
 	return shouldDeploy, runMode, nil
 }
 
-// legacyResourceCollector handles gathering all required resources for a legacy template.
-type legacyResourceCollector struct {
-	template      *appTemplateManifest
-	isInteractive bool
-	// Input values from flags
-	warehouseID     string
-	servingEndpoint string
-	experimentID    string
-	instanceName    string
-	databaseName    string
-	ucVolume        string
-}
-
-// legacyResources holds all collected resource values.
-type legacyResources struct {
-	warehouseID     string
-	servingEndpoint string
-	experimentID    string
-	instanceName    string
-	databaseName    string
-	ucVolume        string
-}
-
-// collectAll gathers all required resources for the template.
-func (c *legacyResourceCollector) collectAll(ctx context.Context) (*legacyResources, error) {
-	resources := &legacyResources{}
-
-	// Get warehouse ID if needed
-	warehouseID, err := getWarehouseIDForTemplate(ctx, c.template, c.warehouseID, c.isInteractive)
-	if err != nil {
-		return nil, err
-	}
-	resources.warehouseID = warehouseID
-
-	// Get serving endpoint if needed
-	servingEndpoint, err := getServingEndpointForTemplate(ctx, c.template, c.servingEndpoint, c.isInteractive)
-	if err != nil {
-		return nil, err
-	}
-	resources.servingEndpoint = servingEndpoint
-
-	// Get experiment ID if needed
-	experimentID, err := getExperimentIDForTemplate(ctx, c.template, c.experimentID, c.isInteractive)
-	if err != nil {
-		return nil, err
-	}
-	resources.experimentID = experimentID
-
-	// Get database resources if needed
-	instanceName, databaseName, err := getDatabaseForTemplate(ctx, c.template, c.instanceName, c.databaseName, c.isInteractive)
-	if err != nil {
-		return nil, err
-	}
-	resources.instanceName = instanceName
-	resources.databaseName = databaseName
-
-	// Get UC volume if needed
-	ucVolume, err := getUCVolumeForTemplate(ctx, c.template, c.ucVolume, c.isInteractive)
-	if err != nil {
-		return nil, err
-	}
-	resources.ucVolume = ucVolume
-
-	return resources, nil
-}
-
-// getResourceForTemplate is a generic function to get a resource value for a template.
-// It checks if the resource is required, uses the provided value if available,
-// prompts in interactive mode, or returns an error in non-interactive mode.
-func getResourceForTemplate(ctx context.Context, tmpl *appTemplateManifest, providedValue string, isInteractive bool, getter resourceGetter) (string, error) {
-	// Check if template requires this resource
-	if !getter.checkRequired(tmpl) {
-		return "", nil
-	}
-
-	// If value was provided via flag, use it
-	if providedValue != "" {
-		return providedValue, nil
-	}
-
-	// In interactive mode, prompt for resource
-	if isInteractive {
-		value, err := getter.promptFunc(ctx)
-		if err != nil {
-			return "", err
-		}
-		return value, nil
-	}
-
-	// Non-interactive mode without value - return error
-	return "", errors.New(getter.errorMessage)
-}
-
-// getWarehouseIDForTemplate ensures a warehouse ID is available if the template requires one.
-func getWarehouseIDForTemplate(ctx context.Context, tmpl *appTemplateManifest, providedWarehouseID string, isInteractive bool) (string, error) {
-	return getResourceForTemplate(ctx, tmpl, providedWarehouseID, isInteractive, resourceGetter{
-		checkRequired: requiresSQLWarehouse,
-		promptFunc:    prompt.PromptForWarehouse,
-		errorMessage:  "template requires a SQL warehouse. Please provide --warehouse-id",
-	})
-}
-
-// getServingEndpointForTemplate ensures a serving endpoint is available if the template requires one.
-func getServingEndpointForTemplate(ctx context.Context, tmpl *appTemplateManifest, providedEndpoint string, isInteractive bool) (string, error) {
-	return getResourceForTemplate(ctx, tmpl, providedEndpoint, isInteractive, resourceGetter{
-		checkRequired: requiresServingEndpoint,
-		promptFunc:    prompt.PromptForServingEndpoint,
-		errorMessage:  "template requires a serving endpoint. Please provide --serving-endpoint",
-	})
-}
-
-// getExperimentIDForTemplate ensures an experiment ID is available if the template requires one.
-func getExperimentIDForTemplate(ctx context.Context, tmpl *appTemplateManifest, providedExperimentID string, isInteractive bool) (string, error) {
-	return getResourceForTemplate(ctx, tmpl, providedExperimentID, isInteractive, resourceGetter{
-		checkRequired: requiresExperiment,
-		promptFunc:    prompt.PromptForExperiment,
-		errorMessage:  "template requires an MLflow experiment. Please provide --experiment-id",
-	})
-}
-
-// getDatabaseForTemplate ensures database instance and name are available if the template requires them.
-// Returns instanceName and databaseName or empty strings if not needed/available.
-func getDatabaseForTemplate(ctx context.Context, tmpl *appTemplateManifest, providedInstanceName, providedDatabaseName string, isInteractive bool) (string, string, error) {
-	// Check if template requires a database
-	if !requiresDatabase(tmpl) {
-		return "", "", nil
-	}
-
-	instanceName := providedInstanceName
-	databaseName := providedDatabaseName
-
-	// In interactive mode, prompt for both if not provided
-	if isInteractive {
-		// Prompt for instance name if not provided
-		if instanceName == "" {
-			var err error
-			instanceName, err = prompt.PromptForDatabaseInstance(ctx)
-			if err != nil {
-				return "", "", err
-			}
-		}
-
-		// Prompt for database name if not provided
-		if databaseName == "" {
-			var err error
-			databaseName, err = prompt.PromptForDatabaseName(ctx, instanceName)
-			if err != nil {
-				return "", "", err
-			}
-		}
-
-		return instanceName, databaseName, nil
-	}
-
-	// Non-interactive mode - both must be provided
-	if instanceName == "" || databaseName == "" {
-		return "", "", errors.New("template requires a database. Please provide both --instance-name and --database-name")
-	}
-
-	return instanceName, databaseName, nil
-}
-
-// getUCVolumeForTemplate ensures a UC volume path is available if the template requires one.
-func getUCVolumeForTemplate(ctx context.Context, tmpl *appTemplateManifest, providedVolume string, isInteractive bool) (string, error) {
-	return getResourceForTemplate(ctx, tmpl, providedVolume, isInteractive, resourceGetter{
-		checkRequired: requiresUCVolume,
-		promptFunc:    prompt.PromptForUCVolume,
-		errorMessage:  "template requires a Unity Catalog volume. Please provide --uc-volume",
-	})
-}
-
-// inlineAppYmlIntoBundle checks for app.yml in the directory, inlines it into databricks.yml, and deletes app.yml.
-func inlineAppYmlIntoBundle(ctx context.Context, dir string) error {
-	// Change to the directory
-	originalDir, err := os.Getwd()
-	if err != nil {
-		return fmt.Errorf("failed to get current directory: %w", err)
-	}
-
-	if err := os.Chdir(dir); err != nil {
-		return fmt.Errorf("failed to change to directory: %w", err)
-	}
-	defer func() {
-		_ = os.Chdir(originalDir)
-	}()
-
-	// Read the databricks.yml file
-	databricksYmlPath := "databricks.yml"
-	databricksData, err := os.ReadFile(databricksYmlPath)
-	if err != nil {
-		return fmt.Errorf("failed to read databricks.yml: %w", err)
-	}
-
-	// Parse databricks.yml as yaml.Node to preserve field names
-	var databricksNode yaml.Node
-	err = yaml.Unmarshal(databricksData, &databricksNode)
-	if err != nil {
-		return fmt.Errorf("failed to parse databricks.yml: %w", err)
-	}
-
-	// Convert yaml.Node to dyn.Value preserving field names
-	configValue, err := yamlNodeToDynValue(&databricksNode)
-	if err != nil {
-		return fmt.Errorf("failed to convert databricks config: %w", err)
-	}
-
-	// Get the app value from resources.apps.app
-	appValue, err := dyn.GetByPath(configValue, dyn.MustPathFromString("resources.apps.app"))
-	if err != nil {
-		return fmt.Errorf("failed to get app from databricks.yml: %w", err)
-	}
-
-	// Inline the app config file (checks for app.yml or app.yaml and inlines if found)
-	appConfigFile, err := inlineAppConfigFile(&appValue)
-	if err != nil {
-		return fmt.Errorf("failed to inline app config: %w", err)
-	}
-
-	// If no app config file was found, nothing to do
-	if appConfigFile == "" {
-		return nil
-	}
-
-	// Set the updated app value back
-	configValue, err = dyn.SetByPath(configValue, dyn.MustPathFromString("resources.apps.app"), appValue)
-	if err != nil {
-		return fmt.Errorf("failed to set updated app value: %w", err)
-	}
-
-	// Extract the top-level map back and set explicit line numbers for ordering
-	configMap, ok := configValue.AsMap()
-	if !ok {
-		return errors.New("config is not a map")
-	}
-
-	// Define the desired order with explicit line numbers
-	keyOrder := map[string]int{
-		"bundle":    1,
-		"workspace": 2,
-		"variables": 3,
-		"resources": 4,
-	}
-
-	updatedConfig := make(map[string]dyn.Value)
-	for _, pair := range configMap.Pairs() {
-		key := pair.Key.MustString()
-		value := pair.Value
-
-		// Set the line number based on the desired order
-		if lineNum, ok := keyOrder[key]; ok {
-			value = dyn.NewValue(value.Value(), []dyn.Location{{Line: lineNum}})
-		}
-
-		updatedConfig[key] = value
-	}
-
-	// Save the updated databricks.yml (force=true since we're updating the file we just created)
-	saver := yamlsaver.NewSaver()
-	err = saver.SaveAsYAML(updatedConfig, databricksYmlPath, true)
-	if err != nil {
-		return fmt.Errorf("failed to save databricks.yml: %w", err)
-	}
-
-	// Add blank lines between top-level keys for better readability
-	err = addBlankLinesBetweenTopLevelKeys(databricksYmlPath)
-	if err != nil {
-		return fmt.Errorf("failed to format databricks.yml: %w", err)
-	}
-
-	// Delete the app config file
-	if err := os.Remove(appConfigFile); err != nil {
-		return fmt.Errorf("failed to remove %s: %w", appConfigFile, err)
-	}
-	cmdio.LogString(ctx, "✓ Inlined and removed "+appConfigFile)
-	return nil
-}
-
-// copyDir recursively copies a directory tree from src to dst.
-func copyDir(src, dst string) error {
-	// Get source directory info
-	srcInfo, err := os.Stat(src)
-	if err != nil {
-		return err
-	}
-
-	// Create destination directory
-	if err := os.MkdirAll(dst, srcInfo.Mode()); err != nil {
-		return err
-	}
-
-	// Read source directory
-	entries, err := os.ReadDir(src)
-	if err != nil {
-		return err
-	}
-
-	for _, entry := range entries {
-		srcPath := filepath.Join(src, entry.Name())
-		dstPath := filepath.Join(dst, entry.Name())
-
-		if entry.IsDir() {
-			// Recursively copy subdirectory
-			if err := copyDir(srcPath, dstPath); err != nil {
-				return err
-			}
-		} else {
-			// Copy file
-			if err := copyFile(srcPath, dstPath); err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-// copyFile copies a single file from src to dst.
-func copyFile(src, dst string) error {
-	srcFile, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-	defer srcFile.Close()
-
-	// Get source file info for permissions
-	srcInfo, err := srcFile.Stat()
-	if err != nil {
-		return err
-	}
-
-	dstFile, err := os.OpenFile(dst, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, srcInfo.Mode())
-	if err != nil {
-		return err
-	}
-	defer dstFile.Close()
-
-	if _, err := srcFile.WriteTo(dstFile); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// runLegacyTemplateInit initializes a project using a legacy template.
-// All resource parameters are optional and will be passed to the template if provided.
-// startCommand from the manifest (if provided) overrides the default command for running the app.
-func runLegacyTemplateInit(ctx context.Context, selectedTemplate *appTemplateManifest, appName, outputDir, warehouseID, servingEndpoint, experimentID, instanceName, databaseName, ucVolume, workspaceHost, profile string, shouldDeploy bool, runMode prompt.RunMode) error {
-	// Determine the destination directory
-	destDir := appName
-	if outputDir != "" {
-		destDir = filepath.Join(outputDir, appName)
-	}
-
-	// Check if directory already exists
-	if _, err := os.Stat(destDir); err == nil {
-		return fmt.Errorf("directory %s already exists", destDir)
-	}
-
-	// Create a temporary directory for cloning the repo
-	tmpDir, err := os.MkdirTemp("", "databricks-app-template-*")
-	if err != nil {
-		return fmt.Errorf("failed to create temp directory: %w", err)
-	}
-	defer os.RemoveAll(tmpDir)
-
-	cmdio.LogString(ctx, "Cloning template repository...")
-
-	// Clone the repository (shallow clone)
-	if err := git.Clone(ctx, selectedTemplate.GitRepo, "", tmpDir); err != nil {
-		return fmt.Errorf("failed to clone repository: %w", err)
-	}
-
-	// Source path is the template directory within the cloned repo
-	srcPath := filepath.Join(tmpDir, selectedTemplate.Path)
-	if _, err := os.Stat(srcPath); err != nil {
-		return fmt.Errorf("template path %s not found in repository: %w", selectedTemplate.Path, err)
-	}
-
-	// Copy the template directory to the destination
-	cmdio.LogString(ctx, fmt.Sprintf("Copying template files to %s...", destDir))
-	if err := copyDir(srcPath, destDir); err != nil {
-		return fmt.Errorf("failed to copy template: %w", err)
-	}
-
-	// Remove .git directory if it exists in the destination
-	gitDir := filepath.Join(destDir, ".git")
-	if _, err := os.Stat(gitDir); err == nil {
-		if err := os.RemoveAll(gitDir); err != nil {
-			return fmt.Errorf("failed to remove .git directory: %w", err)
-		}
-	}
-
-	// Build bundle variables for databricks.yml
-	variablesBuilder := newVariablesBuilder()
-	variablesBuilder.addWarehouse(warehouseID)
-	variablesBuilder.addServingEndpoint(servingEndpoint)
-	variablesBuilder.addExperiment(experimentID)
-	variablesBuilder.addDatabase(instanceName, databaseName)
-	variablesBuilder.addUCVolume(ucVolume)
-
-	// Build resource bindings for databricks.yml
-	bindingsBuilder := newResourceBindingsBuilder()
-	bindingsBuilder.addWarehouse(warehouseID)
-	bindingsBuilder.addServingEndpoint(servingEndpoint)
-	bindingsBuilder.addExperiment(experimentID)
-	bindingsBuilder.addDatabase(instanceName, databaseName)
-
-	// Create databricks.yml using template
-	vars := templateVars{
-		ProjectName:      appName,
-		AppName:          appName,
-		AppDescription:   selectedTemplate.Manifest.Description,
-		WorkspaceHost:    workspaceHost,
-		BundleVariables:  variablesBuilder.build(),
-		ResourceBindings: bindingsBuilder.build(),
-		UserAPIScopes:    selectedTemplate.Manifest.UserAPIScopes,
-	}
-
-	tmpl, err := template.New("databricks.yml").Parse(databricksYmlTemplate)
-	if err != nil {
-		return fmt.Errorf("failed to parse databricks.yml template: %w", err)
-	}
-
-	var databricksYmlBuf bytes.Buffer
-	if err := tmpl.Execute(&databricksYmlBuf, vars); err != nil {
-		return fmt.Errorf("failed to execute databricks.yml template: %w", err)
-	}
-
-	databricksYmlPath := filepath.Join(destDir, "databricks.yml")
-	if err := os.WriteFile(databricksYmlPath, databricksYmlBuf.Bytes(), 0o644); err != nil {
-		return fmt.Errorf("failed to write databricks.yml: %w", err)
-	}
-
-	cmdio.LogString(ctx, "✓ Created databricks.yml")
-
-	// Write .gitignore if it doesn't exist
-	if err := writeGitignoreIfMissing(ctx, destDir); err != nil {
-		// Log warning but don't fail - .gitignore is optional
-		cmdio.LogString(ctx, fmt.Sprintf("⚠ failed to create .gitignore: %v", err))
-	}
-
-	// Generate .env file from app.yml BEFORE inlining (inlining deletes app.yml)
-	if err := generateEnvFileForLegacyTemplate(ctx, destDir, workspaceHost, profile, appName, warehouseID, servingEndpoint, experimentID, instanceName, databaseName, ucVolume); err != nil {
-		// Log warning but don't fail - .env is optional
-		cmdio.LogString(ctx, fmt.Sprintf("⚠ failed to generate .env file: %v", err))
-	}
-
-	// Check for app.yml in the destination directory and inline it into databricks.yml
-	if err := inlineAppYmlIntoBundle(ctx, destDir); err != nil {
-		return fmt.Errorf("failed to inline app.yml: %w", err)
-	}
-
-	// Get absolute path
-	absOutputDir, err := filepath.Abs(destDir)
-	if err != nil {
-		absOutputDir = destDir
-	}
-
-	// Count files in destination directory
-	fileCount := 0
-	err = filepath.Walk(destDir, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-		if !info.IsDir() {
-			fileCount++
-		}
-		return nil
-	})
-	if err != nil {
-		fileCount = 0
-	}
-
-	return runPostCreationSteps(ctx, absOutputDir, appName, fileCount, shouldDeploy, runMode, selectedTemplate.Manifest.StartCommand)
-}
-
-// handleLegacyTemplateInit handles the common logic for initializing a legacy template.
-// It gets the app name, collects resources, determines deploy/run options, and calls runLegacyTemplateInit.
-func handleLegacyTemplateInit(ctx context.Context, legacyTemplate *appTemplateManifest, opts createOptions, isInteractive bool, workspaceHost, profile string) error {
-	// Get app name
-	appName := opts.name
-	if appName == "" {
-		if !isInteractive {
-			return errors.New("--name is required in non-interactive mode")
-		}
-		var err error
-		appName, err = prompt.PromptForProjectName(ctx, opts.outputDir, legacyTemplate.Path)
-		if err != nil {
-			return err
-		}
-	} else {
-		// Validate name in non-interactive mode
-		if err := prompt.ValidateProjectName(appName); err != nil {
-			return err
-		}
-	}
-
-	// Collect all required resources
-	collector := &legacyResourceCollector{
-		template:        legacyTemplate,
-		isInteractive:   isInteractive,
-		warehouseID:     opts.warehouseID,
-		servingEndpoint: opts.servingEndpoint,
-		experimentID:    opts.experimentID,
-		instanceName:    opts.instanceName,
-		databaseName:    opts.databaseName,
-		ucVolume:        opts.ucVolume,
-	}
-	resources, err := collector.collectAll(ctx)
-	if err != nil {
-		return err
-	}
-
-	// Determine deploy and run options
-	deployRun := &deployRunConfig{
-		deploy:        opts.deploy,
-		deployChanged: opts.deployChanged,
-		run:           opts.run,
-		runChanged:    opts.runChanged,
-		isInteractive: isInteractive,
-	}
-	shouldDeploy, runMode, err := deployRun.resolve(ctx)
-	if err != nil {
-		return err
-	}
-
-	return runLegacyTemplateInit(ctx, legacyTemplate, appName, opts.outputDir,
-		resources.warehouseID, resources.servingEndpoint, resources.experimentID,
-		resources.instanceName, resources.databaseName, resources.ucVolume,
-		workspaceHost, profile, shouldDeploy, runMode)
-}
-
 func runCreate(ctx context.Context, opts createOptions) error {
 	var selectedFeatures []string
 	var dependencies map[string]string
@@ -1448,15 +505,15 @@ func runCreate(ctx context.Context, opts createOptions) error {
 	if opts.templatePath != "" {
 		// Check if it's a legacy template identifier (not a URL or local path)
 		if !strings.HasPrefix(opts.templatePath, "https://") && !strings.HasPrefix(opts.templatePath, "/") && !strings.HasPrefix(opts.templatePath, "./") && !strings.HasPrefix(opts.templatePath, "../") {
-			manifests, err := loadLegacyTemplates()
+			templates, err := legacytemplates.LoadLegacyTemplates()
 			if err != nil {
 				return err
 			}
 
 			// Check if the template path matches a legacy template
-			if legacyTemplate := findLegacyTemplateByPath(manifests, opts.templatePath); legacyTemplate != nil {
+			if legacyTemplate := legacytemplates.FindLegacyTemplateByPath(templates, opts.templatePath); legacyTemplate != nil {
 				log.Infof(ctx, "Using legacy template: %s", opts.templatePath)
-				return handleLegacyTemplateInit(ctx, legacyTemplate, opts, isInteractive, workspaceHost, profile)
+				return legacytemplates.HandleLegacyTemplateInit(ctx, legacyTemplate, opts.name, opts.nameProvided, opts.outputDir, opts.warehouseID, opts.servingEndpoint, opts.experimentID, opts.instanceName, opts.databaseName, opts.ucVolume, opts.deploy, opts.deployChanged, opts.run, opts.runChanged, isInteractive, workspaceHost, profile)
 			}
 		}
 	}
@@ -1464,7 +521,7 @@ func runCreate(ctx context.Context, opts createOptions) error {
 	// Step 1: Prompt for template type (AppKit vs Legacy) in interactive mode
 	selectedTemplateType := templateTypeAppKit // default
 	if isInteractive && opts.templatePath == "" {
-		tmplType, err := promptForTemplateType(ctx)
+		tmplType, err := legacytemplates.PromptForTemplateType(ctx)
 		if err != nil {
 			return err
 		}
@@ -1473,17 +530,17 @@ func runCreate(ctx context.Context, opts createOptions) error {
 
 	// If legacy template is selected, use the legacy template flow
 	if selectedTemplateType == templateTypeLegacy {
-		manifests, err := loadLegacyTemplates()
+		templates, err := legacytemplates.LoadLegacyTemplates()
 		if err != nil {
 			return err
 		}
 
-		selectedTemplate, err := promptForLegacyTemplate(ctx, manifests)
+		selectedTemplate, err := legacytemplates.PromptForLegacyTemplate(ctx, templates)
 		if err != nil {
 			return err
 		}
 
-		return handleLegacyTemplateInit(ctx, selectedTemplate, opts, isInteractive, workspaceHost, profile)
+		return legacytemplates.HandleLegacyTemplateInit(ctx, selectedTemplate, opts.name, opts.nameProvided, opts.outputDir, opts.warehouseID, opts.servingEndpoint, opts.experimentID, opts.instanceName, opts.databaseName, opts.ucVolume, opts.deploy, opts.deployChanged, opts.run, opts.runChanged, isInteractive, workspaceHost, profile)
 	}
 
 	// Use features from flags if provided
@@ -1764,11 +821,6 @@ func runCreate(ctx context.Context, opts createOptions) error {
 	})
 	if runErr != nil {
 		return runErr
-	}
-
-	// Write .gitignore if it doesn't exist
-	if err := writeGitignoreIfMissing(ctx, absOutputDir); err != nil {
-		cmdio.LogString(ctx, fmt.Sprintf("⚠ failed to create .gitignore: %v", err))
 	}
 
 	return runPostCreationSteps(ctx, absOutputDir, opts.name, fileCount, shouldDeploy, runMode)
