+++ cmd/apps/init_test.go
@@ -1,14 +1,10 @@
 package apps
 
 import (
-	"bytes"
-	"strings"
 	"testing"
-	"text/template"
 
 	"github.com/databricks/cli/libs/apps/prompt"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestIsTextFile(t *testing.T) {
@@ -244,183 +240,4 @@ func TestParseDeployAndRunFlags(t *testing.T) {
 			assert.Equal(t, tt.wantRunMode, runMode)
 		})
 	}
-}
-
-func TestResourceBindingsBuilder(t *testing.T) {
-	tests := []struct {
-		name     string
-		addOps   func(*resourceBindingsBuilder)
-		expected string
-	}{
-		{
-			name: "warehouse only",
-			addOps: func(b *resourceBindingsBuilder) {
-				b.addWarehouse("abc123")
-			},
-			expected: "        - name: sql-warehouse\n          description: SQL Warehouse for analytics\n          sql_warehouse:\n            id: ${var.warehouse_id}\n            permission: CAN_USE",
-		},
-		{
-			name: "multiple resources",
-			addOps: func(b *resourceBindingsBuilder) {
-				b.addWarehouse("wh1")
-				b.addServingEndpoint("ep1")
-				b.addExperiment("exp1")
-			},
-			expected: "        - name: sql-warehouse\n          description: SQL Warehouse for analytics\n          sql_warehouse:\n            id: ${var.warehouse_id}\n            permission: CAN_USE\n        - name: serving-endpoint\n          description: Model serving endpoint\n          serving_endpoint:\n            name: ${var.serving_endpoint_name}\n            permission: CAN_QUERY\n        - name: experiment\n          description: MLflow experiment\n          experiment:\n            experiment_id: ${var.experiment_id}\n            permission: CAN_MANAGE",
-		},
-		{
-			name: "empty resources",
-			addOps: func(b *resourceBindingsBuilder) {
-				b.addWarehouse("")
-				b.addServingEndpoint("")
-			},
-			expected: "",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			builder := newResourceBindingsBuilder()
-			tt.addOps(builder)
-			result := builder.build()
-			assert.Equal(t, tt.expected, result)
-		})
-	}
-}
-
-func TestHasResourceSpec(t *testing.T) {
-	tests := []struct {
-		name     string
-		manifest *appTemplateManifest
-		checker  resourceSpecChecker
-		expected bool
-	}{
-		{
-			name: "has SQL warehouse",
-			manifest: &appTemplateManifest{
-				Manifest: manifest{
-					ResourceSpecs: []resourceSpec{
-						{Name: "warehouse", SQLWarehouseSpec: &map[string]any{"id": "abc"}},
-					},
-				},
-			},
-			checker:  func(s *resourceSpec) bool { return s.SQLWarehouseSpec != nil },
-			expected: true,
-		},
-		{
-			name: "no SQL warehouse",
-			manifest: &appTemplateManifest{
-				Manifest: manifest{
-					ResourceSpecs: []resourceSpec{
-						{Name: "endpoint", ServingEndpointSpec: &map[string]any{"name": "ep"}},
-					},
-				},
-			},
-			checker:  func(s *resourceSpec) bool { return s.SQLWarehouseSpec != nil },
-			expected: false,
-		},
-		{
-			name: "empty specs",
-			manifest: &appTemplateManifest{
-				Manifest: manifest{
-					ResourceSpecs: []resourceSpec{},
-				},
-			},
-			checker:  func(s *resourceSpec) bool { return s.SQLWarehouseSpec != nil },
-			expected: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := hasResourceSpec(tt.manifest, tt.checker)
-			assert.Equal(t, tt.expected, result)
-		})
-	}
-}
-
-func TestDatabricksYmlTemplateWithUserAPIScopes(t *testing.T) {
-	tests := []struct {
-		name          string
-		vars          templateVars
-		expectScopes  bool
-		expectedScope string
-	}{
-		{
-			name: "with user_api_scopes",
-			vars: templateVars{
-				ProjectName:    "test-app",
-				AppName:        "test-app",
-				AppDescription: "Test application",
-				UserAPIScopes:  []string{"sql", "catalog.connections"},
-			},
-			expectScopes:  true,
-			expectedScope: "user_api_scopes:\n        - sql\n        - catalog.connections",
-		},
-		{
-			name: "without user_api_scopes",
-			vars: templateVars{
-				ProjectName:    "test-app",
-				AppName:        "test-app",
-				AppDescription: "Test application",
-				UserAPIScopes:  nil,
-			},
-			expectScopes: false,
-		},
-		{
-			name: "with empty user_api_scopes",
-			vars: templateVars{
-				ProjectName:    "test-app",
-				AppName:        "test-app",
-				AppDescription: "Test application",
-				UserAPIScopes:  []string{},
-			},
-			expectScopes: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tmpl, err := template.New("databricks.yml").Parse(databricksYmlTemplate)
-			require.NoError(t, err)
-
-			var buf bytes.Buffer
-			err = tmpl.Execute(&buf, tt.vars)
-			require.NoError(t, err)
-
-			output := buf.String()
-
-			if tt.expectScopes {
-				assert.Contains(t, output, tt.expectedScope)
-			} else {
-				assert.NotContains(t, output, "user_api_scopes")
-			}
-
-			// Verify basic structure is present
-			assert.Contains(t, output, "bundle:")
-			assert.Contains(t, output, "name: test-app")
-			assert.Contains(t, output, "resources:")
-			assert.Contains(t, output, "apps:")
-		})
-	}
-}
-
-func TestUserAPIScopesInLegacyManifest(t *testing.T) {
-	// Verify that the manifest JSON includes templates with user_api_scopes
-	manifests, err := loadLegacyTemplates()
-	require.NoError(t, err)
-	require.NotNil(t, manifests)
-
-	// Find a template with user_api_scopes (e.g., dash-data-app-obo-user)
-	var foundWithScopes bool
-	for i := range manifests.Templates {
-		tmpl := &manifests.Templates[i]
-		if strings.Contains(tmpl.Path, "obo-user") {
-			foundWithScopes = true
-			assert.NotEmpty(t, tmpl.Manifest.UserAPIScopes, "Template %s should have user_api_scopes", tmpl.Path)
-			assert.Contains(t, tmpl.Manifest.UserAPIScopes, "sql", "Template %s should include 'sql' scope", tmpl.Path)
-		}
-	}
-
-	assert.True(t, foundWithScopes, "Should have found at least one template with user_api_scopes")
 }
