+++ cmd/apps/internal/yamlutil/yaml_utils_test.go
@@ -1,4 +1,4 @@
-package apps
+package yamlutil
 
 import (
 	"testing"
