+++ cmd/apps/legacytemplates/app-template-app-manifests.json
