+++ cmd/apps/legacytemplates/env_builder_test.go
@@ -1,4 +1,4 @@
-package apps
+package legacytemplates
 
 import (
 	"os"
