+++ cmd/apps/legacytemplates/gitignore.tmpl
